(WIP) Possibility to detach swift from controllers
This change allows for the overriding of the use of primary-controller
for the swift deployment.  The following new heira values can be used
to install swift on nodes other than controllers.

*) swift_master_role - The role used to designate the primary swift
                       node. Defaults to 'primary-controller'.
*) swift_nodes - The list of swift nodes that will be configured.
                 Defaults to to using the hiera('controllers')

One item not addressed is the use of the heira setting
'primary_controller' for determining the primary proxy setting.

Partial blueprint: detach-components-from-controllers

Change-Id: I8fe388f59892c05ff0fa55e06843fe0cd833c34c